requirements.txt
pandas
requests
sqlalchemy
python-dotenv
jupyter
apache-airflow
